+++ src/accelerate/utils/launch.py
@@ -58,7 +58,7 @@ def _get_mpirun_args():
     # Find the MPI program name
     mpi_apps = [x for x in ["mpirun", "mpiexec"] if which(x)]
 
-    if not mpi_apps:
+    if len(mpi_apps) == 0:
         raise OSError("mpirun or mpiexec were not found. Ensure that Intel MPI, Open MPI, or MVAPICH are installed.")
 
     # Call the app with the --version flag to determine which MPI app is installed
@@ -80,9 +80,7 @@ def prepare_simple_launcher_cmd_env(args: argparse.Namespace) -> Tuple[List[str]
     if args.no_python and args.module:
         raise ValueError("--module and --no_python cannot be used together")
 
-    mpirun_hostfile = getattr(args, "mpirun_hostfile", None)
-
-    if mpirun_hostfile:
+    if args.mpirun_hostfile is not None:
         mpi_app_name, hostfile_arg, num_proc_arg, proc_per_node_arg = _get_mpirun_args()
         mpirun_ccl = getattr(args, "mpirun_ccl", None)
         num_machines = args.num_machines
@@ -113,7 +111,7 @@ def prepare_simple_launcher_cmd_env(args: argparse.Namespace) -> Tuple[List[str]
         current_env["MASTER_ADDR"] = args.main_process_ip
         current_env["MASTER_PORT"] = str(args.main_process_port)
 
-        if mpirun_hostfile:
+        if args.mpirun_hostfile is not None:
             current_env["CCL_WORKER_COUNT"] = mpirun_ccl
     elif args.num_processes > 1:
         current_env["MASTER_ADDR"] = args.main_process_ip if args.main_process_ip is not None else "127.0.0.1"
